Extract toolbar stuff into use hooks
