+++ src/renderer/src/components/Toolbar.tsx
@@ -1,47 +1,17 @@
 import { useState, useRef, useEffect } from "react";
-import { v4 as uuid } from "uuid";
 import { PenLine, Moon, Sun } from "lucide-react";
 import TerraForgeLogotype from "../assets/terraForgeLogotype.svg?react";
 import { useMachineStore } from "../store/machineStore";
 import { useCanvasStore } from "../store/canvasStore";
-import { useTaskStore } from "../store/taskStore";
 import { useThemeStore } from "../store/themeStore";
-import {
-  type VectorObject,
-  type MachineConfig,
-  type GcodeOptions,
-  type SvgImport,
-  type SvgPath,
-  type SvgLayer,
-  type CanvasLayout,
-  type PageTemplate,
-  DEFAULT_HATCH_SPACING_MM,
-  DEFAULT_HATCH_ANGLE_DEG,
-} from "../../../types";
 import { MachineConfigDialog } from "./MachineConfigDialog";
 import { GcodeOptionsDialog } from "./GcodeOptionsDialog";
 import { CloseLayoutDialog } from "./CloseLayoutDialog";
 import { ConfirmDialog } from "./ConfirmDialog";
 import { AboutDialog } from "./AboutDialog";
-import type { GcodePrefs } from "./GcodeOptionsDialog";
-import {
-  getAccumulatedTransform,
-  applyMatrixToPathD,
-  computePathsBounds,
-} from "../utils/svgTransform";
-import { generateHatchPaths } from "../utils/hatchFill";
-import { parseGcode } from "../utils/gcodeParser";
-import { importPdf } from "../utils/pdfImport";
-import {
-  findContainingLayerId,
-  getEffectiveFill,
-  getLayerName,
-  hasVisibleStroke,
-  isDisplayNone,
-  isLayerGroup,
-  parseSvgLengthMM,
-  shapeToPathD,
-} from "../features/imports/services/svgImportHelpers";
+import { useImportActions } from "../features/imports/hooks/useImportActions";
+import { useLayoutActions } from "../features/layout/hooks/useLayoutActions";
+import { useJobActions } from "../features/machine/hooks/useJobActions";
 
 interface ToolbarProps {
   showJog?: boolean;
@@ -55,27 +25,13 @@ export function Toolbar({
   const configs = useMachineStore((s) => s.configs);
   const activeConfigId = useMachineStore((s) => s.activeConfigId);
   const setActiveConfigId = useMachineStore((s) => s.setActiveConfigId);
-  const activeConfig = useMachineStore((s) => s.activeConfig);
   const connected = useMachineStore((s) => s.connected);
   const wsLive = useMachineStore((s) => s.wsLive);
   const fwInfo = useMachineStore((s) => s.fwInfo);
-  const setConnected = useMachineStore((s) => s.setConnected);
-  const setSelectedJobFile = useMachineStore((s) => s.setSelectedJobFile);
+
   const imports = useCanvasStore((s) => s.imports);
   const selectedImportId = useCanvasStore((s) => s.selectedImportId);
-  const addImport = useCanvasStore((s) => s.addImport);
   const clearImports = useCanvasStore((s) => s.clearImports);
-  const loadLayout = useCanvasStore((s) => s.loadLayout);
-  const layerGroups = useCanvasStore((s) => s.layerGroups);
-  const toVectorObjectsForGroup = useCanvasStore(
-    (s) => s.toVectorObjectsForGroup,
-  );
-  const toVectorObjectsUngrouped = useCanvasStore(
-    (s) => s.toVectorObjectsUngrouped,
-  );
-  const setGcodeToolpath = useCanvasStore((s) => s.setGcodeToolpath);
-  const setGcodeSource = useCanvasStore((s) => s.setGcodeSource);
-  const selectToolpath = useCanvasStore((s) => s.selectToolpath);
   const copyImport = useCanvasStore((s) => s.copyImport);
   const cutImport = useCanvasStore((s) => s.cutImport);
   const pasteImport = useCanvasStore((s) => s.pasteImport);
@@ -88,562 +44,40 @@ export function Toolbar({
   const setPageTemplate = useCanvasStore((s) => s.setPageTemplate);
   const pageSizes = useCanvasStore((s) => s.pageSizes);
   const setPageSizes = useCanvasStore((s) => s.setPageSizes);
-  const upsertTask = useTaskStore((s) => s.upsertTask);
-  const registerCancelCallback = useTaskStore((s) => s.registerCancelCallback);
-  const unregisterCancelCallback = useTaskStore(
-    (s) => s.unregisterCancelCallback,
-  );
+
   const theme = useThemeStore((s) => s.theme);
   const toggleTheme = useThemeStore((s) => s.toggleTheme);
 
-  // Holds a reference to the active G-code worker so cancellation can reach it.
-  const activeWorkerRef = useRef<{ worker: Worker; taskId: string } | null>(
-    null,
-  );
-
   const [showGcodeDialog, setShowGcodeDialog] = useState(false);
-  const [isConnecting, setIsConnecting] = useState(false);
-  const [generating, setGenerating] = useState(false);
   const [showSettings, setShowSettings] = useState(false);
-  const [showCloseDialog, setShowCloseDialog] = useState(false);
   const [showAbout, setShowAbout] = useState(false);
-  /** Parsed layout waiting for the user to confirm overwriting the canvas. */
-  const [pendingLayout, setPendingLayout] = useState<CanvasLayout | null>(null);
 
-  const handleConnect = async () => {
-    const cfg = activeConfig();
-    if (!cfg) return;
-    const taskId = uuid();
-    const label =
-      cfg.connection.type === "wifi"
-        ? `Connecting to ${cfg.connection.host}…`
-        : `Connecting to ${cfg.connection.serialPath}…`;
-    upsertTask({
-      id: taskId,
-      type: "ws-connect",
-      label,
-      progress: null,
-      status: "running",
-    });
-    setIsConnecting(true);
-    try {
-      if (cfg.connection.type === "wifi") {
-        await window.terraForge.fluidnc.connectWebSocket(
-          cfg.connection.host!,
-          cfg.connection.port ?? 80,
-          cfg.connection.wsPort,
-        );
-      } else {
-        await window.terraForge.serial.connect(
-          cfg.connection.serialPath!,
-          115200,
-        );
-      }
-      setConnected(true);
-      upsertTask({
-        id: taskId,
-        type: "ws-connect",
-        label: "Connected",
-        progress: 100,
-        status: "completed",
-      });
-    } catch (err) {
-      console.error("Connection failed", err);
-      upsertTask({
-        id: taskId,
-        type: "ws-connect",
-        label: "Connection failed",
-        progress: null,
-        status: "error",
-        error: err instanceof Error ? err.message : String(err),
-      });
-    } finally {
-      setIsConnecting(false);
-    }
-  };
+  // ── Extracted action hooks ────────────────────────────────────────────────
+  const { handleImport } = useImportActions();
 
-  const handleDisconnect = async () => {
-    const cfg = activeConfig();
-    if (!cfg) return;
-    if (cfg.connection.type === "wifi") {
-      await window.terraForge.fluidnc.disconnectWebSocket();
-    } else {
-      await window.terraForge.serial.disconnect();
-    }
-    setConnected(false);
-  };
+  const {
+    handleSaveLayout,
+    handleLoadLayout,
+    handleCloseLayout,
+    doCloseLayout,
+    saveLayoutRef,
+    loadLayoutRef,
+    closeLayoutRef,
+    showCloseDialog,
+    setShowCloseDialog,
+    pendingLayout,
+    setPendingLayout,
+  } = useLayoutActions();
 
-  // ── Unified import handler — routes by file extension ─────────────────────
-  const handleImport = async () => {
-    const filePath = await window.terraForge.fs.openImportDialog();
-    if (!filePath) return;
-    const ext = filePath.split(".").pop()?.toLowerCase() ?? "";
-    if (ext === "svg") {
-      await handleImportSvgFile(filePath);
-    } else if (ext === "pdf") {
-      await handleImportPdfFile(filePath);
-    } else {
-      await handleImportGcodeFile(filePath);
-    }
-  };
+  const {
+    handleConnect,
+    handleDisconnect,
+    isConnecting,
+    handleGenerateGcode,
+    generating,
+  } = useJobActions();
 
-  const handleImportSvgFile = async (filePath: string) => {
-    const taskId = uuid();
-    upsertTask({
-      id: taskId,
-      type: "svg-parse",
-      label: "Parsing SVG…",
-      progress: null,
-      status: "running",
-    });
-
-    try {
-      const raw = await window.terraForge.fs.readFile(filePath);
-      const parser = new DOMParser();
-      const doc = parser.parseFromString(raw, "image/svg+xml");
-
-      const svgEl = doc.querySelector("svg");
-      const vbParts = svgEl
-        ?.getAttribute("viewBox")
-        ?.trim()
-        .split(/[\s,]+/)
-        .map(Number);
-      const viewBoxX = vbParts?.[0] ?? 0;
-      const viewBoxY = vbParts?.[1] ?? 0;
-      const svgWidth = vbParts?.[2] ?? +(svgEl?.getAttribute("width") ?? 100);
-      const svgHeight = vbParts?.[3] ?? +(svgEl?.getAttribute("height") ?? 100);
-
-      // Compute initial scale: mm per SVG user-unit.
-      // SVG width/height attrs carry the physical size (may have mm/cm/in/px units).
-      // If they're in mm the result is exact; px/unitless uses 96 DPI.
-      const physW = parseSvgLengthMM(svgEl?.getAttribute("width"));
-      const physH = parseSvgLengthMM(svgEl?.getAttribute("height"));
-      // Prefer width-based scale; fall back to height, then to the SVG/CSS default
-      // of 1 user unit = 1 CSS pixel at 96dpi (25.4/96 mm/px).  Using 1 mm/unit
-      // here is wrong for percentage-dimensioned SVGs and makes content invisible.
-      const initScale =
-        physW != null
-          ? physW / svgWidth
-          : physH != null
-            ? physH / svgHeight
-            : 25.4 / 96;
-
-      // Default name = filename without extension
-      const name =
-        filePath
-          .split(/[\\/]/)
-          .pop()
-          ?.replace(/\.[^.]+$/, "") ?? "import";
-
-      // ── Detect logical layer groups ─────────────────────────────────────
-      // A <g> element is treated as a layer group when it carries an explicit
-      // `display` declaration in its inline style (e.g. Inkscape sub-layers)
-      // and is not inside a non-rendering context (defs, clipPath, mask…).
-      const layerGroupEls = Array.from(doc.querySelectorAll("g")).filter(
-        (g) => !g.closest("defs, clipPath, mask, symbol") && isLayerGroup(g),
-      );
-      const layers: SvgLayer[] = layerGroupEls.map((g, i) => ({
-        id: g.id || `layer_${i}`,
-        name: getLayerName(g, i),
-        visible: !isDisplayNone(g),
-      }));
-      // Ensure every layer group has an id so findContainingLayerId can match it.
-      layerGroupEls.forEach((g, i) => {
-        if (!g.id) g.id = `layer_${i}`;
-      });
-      const layerGroupIds = new Set(layers.map((l) => l.id));
-
-      // Collect all shape elements, skipping only those inside hidden ancestors
-      // that are NOT tracked layer groups.  Paths inside hidden layer groups
-      // are still collected — their visibility is managed per-layer in the UI.
-      const els = Array.from(
-        doc.querySelectorAll(
-          "path, rect, circle, ellipse, line, polyline, polygon",
-        ),
-      ).filter((el) => {
-        let cur = el.parentElement;
-        while (cur && cur.tagName.toLowerCase() !== "svg") {
-          if (isDisplayNone(cur)) {
-            // Hidden layer group — allowed; layer visibility controls it.
-            if (cur.id && layerGroupIds.has(cur.id)) {
-              cur = cur.parentElement;
-              continue;
-            }
-            return false; // hidden non-layer ancestor — skip
-          }
-          cur = cur.parentElement;
-        }
-        return true;
-      });
-
-      // Track which elements have a visible fill so we can generate hatch lines
-      // for them after coordinates are normalized.
-      const fillFlags: boolean[] = [];
-
-      const paths: SvgPath[] = els.flatMap((el): SvgPath[] => {
-        const rawD = shapeToPathD(el);
-        if (!rawD) return [];
-        // Resolve all ancestor transform attributes and bake them into the
-        // path coordinates so the canvas and G-code worker see pre-transformed
-        // positions (fixes Inkscape layer/group transforms).
-        const matrix = getAccumulatedTransform(el);
-        const d = applyMatrixToPathD(rawD, matrix);
-
-        // Skip paths whose bounding box has no intersection with the SVG
-        // viewport (viewBox rectangle in root user-unit space).  This mirrors
-        // the browser's default overflow:hidden clipping at the SVG boundary,
-        // so terraForge shows exactly the same content as a browser does.
-        const pathBounds = computePathsBounds([d]);
-        if (pathBounds) {
-          const vbMaxX = viewBoxX + svgWidth;
-          const vbMaxY = viewBoxY + svgHeight;
-          if (
-            pathBounds.minX > vbMaxX ||
-            pathBounds.maxX < viewBoxX ||
-            pathBounds.minY > vbMaxY ||
-            pathBounds.maxY < viewBoxY
-          ) {
-            return [];
-          }
-        }
-
-        fillFlags.push(getEffectiveFill(el) !== null);
-        const hasFill = fillFlags[fillFlags.length - 1];
-        const outlineVisible = hasVisibleStroke(el);
-        const tag = el.tagName.toLowerCase();
-        const pathIndex = fillFlags.length; // 1-based after push
-
-        // Best-effort display name: Inkscape label > <title> child > element id >
-        // tagname_N fallback.
-        const inkLabel =
-          el.getAttribute("inkscape:label") ??
-          el.getAttributeNS(
-            "http://www.inkscape.org/namespaces/inkscape",
-            "label",
-          );
-        const titleText = el
-          .querySelector(":scope > title")
-          ?.textContent?.trim();
-        const ownId =
-          el.id && !/^[a-f0-9-]{36}$/.test(el.id) ? el.id : undefined;
-        const label =
-          inkLabel?.trim() || titleText || ownId || `${tag}_${pathIndex}`;
-
-        return [
-          {
-            id: uuid(),
-            d,
-            svgSource: el.outerHTML,
-            visible: true,
-            hasFill,
-            outlineVisible,
-            label,
-            layer: findContainingLayerId(el, layerGroupIds),
-          },
-        ];
-      });
-
-      if (paths.length === 0) {
-        upsertTask({
-          id: taskId,
-          type: "svg-parse",
-          label: "No paths found in SVG",
-          progress: null,
-          status: "error",
-        });
-        return;
-      }
-
-      // After baking transforms the paths live in the transformed coordinate
-      // space. Recompute the content extents so the selection bounding box,
-      // canvas group transform, and G-code origin all reflect the actual
-      // occupied area rather than the raw SVG viewBox.
-      //
-      // We also normalize path coordinates to start at (0, 0) by translating
-      // every path by (-minX, -minY). This ensures the G-code worker always
-      // receives coordinates whose origin is the top-left of the content
-      // bounding box, which is what its transformPt formula expects.
-      const bounds = computePathsBounds(paths.map((p) => p.d));
-      const effW = bounds ? bounds.maxX - bounds.minX : svgWidth;
-      const effH = bounds ? bounds.maxY - bounds.minY : svgHeight;
-
-      const normalizedPaths = bounds
-        ? paths.map((p) => ({
-            ...p,
-            d: applyMatrixToPathD(
-              p.d,
-              new DOMMatrix([1, 0, 0, 1, -bounds.minX, -bounds.minY]),
-            ),
-          }))
-        : paths;
-
-      // Generate hatch-fill lines for shapes that had a visible fill and embed
-      // them on their parent SvgPath so they toggle as a unit.
-      // Spacing is converted from mm to SVG user units using initScale.
-      let finalPaths = normalizedPaths;
-      if (initScale > 0) {
-        const spacingUnits = DEFAULT_HATCH_SPACING_MM / initScale;
-        finalPaths = normalizedPaths.map((np, i) => {
-          if (!fillFlags[i]) return np;
-          const hatchLines = generateHatchPaths(
-            np.d,
-            spacingUnits,
-            DEFAULT_HATCH_ANGLE_DEG,
-          );
-          return hatchLines.length ? { ...np, hatchLines } : np;
-        });
-      }
-
-      const imp: SvgImport = {
-        id: uuid(),
-        name,
-        paths: finalPaths,
-        x: 0,
-        y: 0,
-        scale: initScale,
-        rotation: 0,
-        visible: true,
-        svgWidth: effW,
-        svgHeight: effH,
-        viewBoxX: 0,
-        viewBoxY: 0,
-        hatchEnabled: true,
-        hatchSpacingMM: DEFAULT_HATCH_SPACING_MM,
-        hatchAngleDeg: DEFAULT_HATCH_ANGLE_DEG,
-        layers: layers.length > 0 ? layers : undefined,
-      };
-
-      addImport(imp);
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "SVG imported",
-        progress: 100,
-        status: "completed",
-      });
-    } catch (err) {
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "SVG import failed",
-        progress: null,
-        status: "error",
-        error: String(err),
-      });
-    }
-  };
-
-  const handleImportPdfFile = async (filePath: string) => {
-    const name =
-      filePath
-        .split(/[\\/]/)
-        .pop()
-        ?.replace(/\.[^.]+$/i, "") ?? "import";
-
-    const taskId = uuid();
-    upsertTask({
-      id: taskId,
-      type: "svg-parse",
-      label: `Importing ${name}.pdf…`,
-      progress: null,
-      status: "running",
-    });
-
-    try {
-      const data = await window.terraForge.fs.readFileBinary(filePath);
-      const pdfImports = await importPdf(data, name);
-
-      if (pdfImports.length === 0) {
-        upsertTask({
-          id: taskId,
-          type: "svg-parse",
-          label: "No vector paths found in PDF",
-          progress: null,
-          status: "error",
-        });
-        return;
-      }
-
-      for (const imp of pdfImports) {
-        addImport(imp);
-      }
-
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label:
-          pdfImports.length === 1
-            ? `PDF imported: ${pdfImports[0].name}`
-            : `PDF imported: ${pdfImports.length} pages`,
-        progress: 100,
-        status: "completed",
-      });
-    } catch (err) {
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "PDF import failed",
-        progress: null,
-        status: "error",
-        error: String(err),
-      });
-    }
-  };
-
-  const handleImportGcodeFile = async (filePath: string) => {
-    const name = filePath.split(/[\\/]/).pop() ?? "import.gcode";
-    const taskId = uuid();
-    upsertTask({
-      id: taskId,
-      type: "svg-parse",
-      label: `Importing ${name}…`,
-      progress: null,
-      status: "running",
-    });
-
-    try {
-      const text = await window.terraForge.fs.readFile(filePath);
-      const toolpath = parseGcode(text);
-      setGcodeToolpath(toolpath);
-      setGcodeSource({ path: filePath, name, source: "local" });
-      // Auto-select the toolpath so the canvas, Properties panel, and
-      // Job panel are all in sync from the moment of import.
-      selectToolpath(true);
-      // Selecting this as the queued job (local source — will upload on Start)
-      setSelectedJobFile({ path: filePath, source: "local", name });
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: `G-code imported: ${name}`,
-        progress: 100,
-        status: "completed",
-      });
-    } catch (err) {
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "G-code import failed",
-        progress: null,
-        status: "error",
-        error: String(err),
-      });
-    }
-  };
-
-  const handleSaveLayout = async () => {
-    if (imports.length === 0) return;
-    const taskId = uuid();
-    const baseName = imports.length === 1 ? imports[0].name : "layout";
-    const defaultFilename = `${baseName}.tforge`;
-    const savePath =
-      await window.terraForge.fs.saveLayoutDialog(defaultFilename);
-    if (!savePath) return;
-    upsertTask({
-      id: taskId,
-      type: "svg-parse",
-      label: "Saving layout…",
-      progress: null,
-      status: "running",
-    });
-    try {
-      const layout: CanvasLayout = {
-        tfVersion: 1,
-        savedAt: new Date().toISOString(),
-        imports,
-        layerGroups,
-        pageTemplate,
-      };
-      await window.terraForge.fs.writeFile(
-        savePath,
-        JSON.stringify(layout, null, 2),
-      );
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "Layout saved",
-        progress: 100,
-        status: "completed",
-      });
-    } catch (err) {
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "Save failed",
-        progress: null,
-        status: "error",
-        error: String(err),
-      });
-    }
-  };
-
-  const handleLoadLayout = async () => {
-    const filePath = await window.terraForge.fs.openLayoutDialog();
-    if (!filePath) return;
-    const taskId = uuid();
-    upsertTask({
-      id: taskId,
-      type: "svg-parse",
-      label: "Loading layout…",
-      progress: null,
-      status: "running",
-    });
-    try {
-      const raw = await window.terraForge.fs.readFile(filePath);
-      let layout: CanvasLayout;
-      try {
-        layout = JSON.parse(raw) as CanvasLayout;
-      } catch {
-        throw new Error("Not a valid terraForge layout file.");
-      }
-      if (!Array.isArray(layout.imports)) {
-        throw new Error("Layout file does not contain a valid imports array.");
-      }
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "Layout ready",
-        progress: 100,
-        status: "completed",
-      });
-      if (imports.length > 0) {
-        // Canvas already has content — ask before overwriting.
-        setPendingLayout(layout);
-      } else {
-        loadLayout(layout.imports, layout.layerGroups, layout.pageTemplate);
-      }
-    } catch (err) {
-      upsertTask({
-        id: taskId,
-        type: "svg-parse",
-        label: "Load layout failed",
-        progress: null,
-        status: "error",
-        error: String(err),
-      });
-    }
-  };
-
-  const handleCloseLayout = () => {
-    // Nothing to close.
-    if (imports.length === 0) return;
-    setShowCloseDialog(true);
-  };
-
-  const doCloseLayout = () => {
-    clearImports();
-    setShowCloseDialog(false);
-  };
-
-  // Keep refs current so the menu IPC listeners (subscribed once on mount)
-  // always call the latest function closures without stale state.
-  const saveLayoutRef = useRef(handleSaveLayout);
-  const loadLayoutRef = useRef(handleLoadLayout);
-  const closeLayoutRef = useRef(handleCloseLayout);
-  saveLayoutRef.current = handleSaveLayout;
-  loadLayoutRef.current = handleLoadLayout;
-  closeLayoutRef.current = handleCloseLayout;
-
-  // ── Edit clipboard — keep refs current for use in stable IPC/keyboard listeners ─
+  // ── Edit clipboard — keep refs current for stable IPC/keyboard listeners ─
   const selectedImportIdRef = useRef(selectedImportId);
   const clipboardImportRef = useRef(clipboardImport);
   const allImportsSelectedRef = useRef(allImportsSelected);
@@ -803,340 +237,6 @@ export function Toolbar({
   useEffect(() => {
     window.terraForge.edit.setHasSelection(selectedImportId !== null);
   }, [selectedImportId]);
-
-  const handleGenerateGcode = async (prefs: GcodePrefs) => {
-    const cfg = activeConfig();
-    if (!cfg || imports.length === 0) return;
-
-    // Route to per-group export when the option is enabled and groups exist.
-    if (prefs.exportPerGroup && layerGroups.length > 0) {
-      await handleGenerateGcodePerGroup(prefs, cfg);
-      return;
-    }
-
-    setGenerating(true);
-    const taskId = uuid();
-    const activePageSize = pageTemplate
-      ? pageSizes.find((ps) => ps.id === pageTemplate.sizeId)
-      : undefined;
-    const options: GcodeOptions = {
-      arcFitting: false,
-      arcTolerance: 0.01,
-      optimisePaths: prefs.optimise,
-      joinPaths: prefs.joinPaths,
-      joinTolerance: prefs.joinTolerance,
-      liftPenAtEnd: prefs.liftPenAtEnd,
-      returnToHome: prefs.returnToHome,
-      customStartGcode: prefs.customStartGcode,
-      customEndGcode: prefs.customEndGcode,
-      pageClip:
-        activePageSize && prefs.clipMode !== "none"
-          ? {
-              widthMM: pageTemplate!.landscape
-                ? activePageSize.heightMM
-                : activePageSize.widthMM,
-              heightMM: pageTemplate!.landscape
-                ? activePageSize.widthMM
-                : activePageSize.heightMM,
-              marginMM:
-                prefs.clipMode === "margin"
-                  ? (pageTemplate!.marginMM ?? 20)
-                  : (prefs.clipOffsetMM ?? 0),
-            }
-          : undefined,
-    };
-
-    const worker = new Worker(
-      new URL("../../../workers/svgWorker.ts", import.meta.url),
-      { type: "module" },
-    );
-
-    activeWorkerRef.current = { worker, taskId };
-
-    // Register a direct cancel callback so TaskBar can stop the worker without
-    // an IPC round-trip through the main process.
-    registerCancelCallback(taskId, () => {
-      worker.postMessage({ type: "cancel", taskId });
-    });
-
-    upsertTask({
-      id: taskId,
-      type: "gcode-generate",
-      label: "Generating G-code",
-      progress: 0,
-      status: "running",
-    });
-
-    worker.onmessage = async (e) => {
-      const msg = e.data;
-      if (msg.type === "progress") {
-        upsertTask({
-          id: taskId,
-          type: "gcode-generate",
-          label: "Generating G-code",
-          progress: msg.percent,
-          status: "running",
-        });
-      } else if (msg.type === "complete") {
-        worker.terminate();
-        activeWorkerRef.current = null;
-        unregisterCancelCallback(taskId);
-        setGenerating(false);
-
-        // Build a meaningful default filename from the import names.
-        const baseName =
-          imports.length === 1
-            ? imports[0].name
-            : `${imports[0].name}+${imports.length - 1}`;
-        const safeName = baseName
-          .replace(/\.[^.]+$/, "") // strip any existing extension
-          .replace(/[\\/:*?"<>|]/g, "_"); // remove filesystem-illegal chars
-        const defaultFilename = prefs.optimise
-          ? `${safeName}_opt.gcode`
-          : `${safeName}.gcode`;
-
-        upsertTask({
-          id: taskId,
-          type: "gcode-generate",
-          label: "G-code ready",
-          progress: 100,
-          status: "completed",
-        });
-
-        // ── Load toolpath preview into canvas (enables tracing during job) ──
-        const toolpath = parseGcode(msg.gcode);
-        setGcodeToolpath(toolpath);
-        selectToolpath(true);
-
-        // ── Upload to SD card (if opted in and connected) ─────────────────
-        if (prefs.uploadToSd && useMachineStore.getState().connected) {
-          const uploadTaskId = uuid();
-          const remotePath = "/" + defaultFilename;
-          try {
-            await window.terraForge.fluidnc.uploadGcode(
-              uploadTaskId,
-              msg.gcode,
-              remotePath,
-            );
-            // Auto-select the uploaded file so Start Job is ready immediately
-            setSelectedJobFile({
-              path: remotePath,
-              source: "sd",
-              name: defaultFilename,
-            });
-            setGcodeSource({
-              path: remotePath,
-              name: defaultFilename,
-              source: "sd",
-            });
-          } catch {
-            // Upload error is already surfaced via the upload task toast
-          }
-        }
-
-        // ── Save to local computer (if opted in) ──────────────────────────
-        if (prefs.saveLocally) {
-          const savePath =
-            await window.terraForge.fs.saveGcodeDialog(defaultFilename);
-          if (savePath) {
-            await window.terraForge.fs.writeFile(savePath, msg.gcode);
-            setGcodeSource({
-              path: savePath,
-              name: defaultFilename,
-              source: "local",
-            });
-          }
-        }
-      } else if (msg.type === "cancelled") {
-        upsertTask({
-          id: taskId,
-          type: "gcode-generate",
-          label: "G-code cancelled",
-          progress: null,
-          status: "cancelled",
-        });
-        worker.terminate();
-        activeWorkerRef.current = null;
-        unregisterCancelCallback(taskId);
-        setGenerating(false);
-      } else if (msg.type === "error") {
-        upsertTask({
-          id: taskId,
-          type: "gcode-generate",
-          label: "G-code failed",
-          progress: null,
-          status: "error",
-          error: msg.error,
-        });
-        worker.terminate();
-        activeWorkerRef.current = null;
-        unregisterCancelCallback(taskId);
-        setGenerating(false);
-      }
-    };
-
-    worker.postMessage({
-      type: "generate",
-      taskId,
-      objects: useCanvasStore.getState().toVectorObjects(),
-      config: cfg,
-      options,
-    });
-  };
-
-  // ── Per-group G-code export ───────────────────────────────────────────────
-  const handleGenerateGcodePerGroup = async (
-    prefs: GcodePrefs,
-    cfg: ReturnType<typeof activeConfig>,
-  ) => {
-    if (!cfg) return;
-    const activePageSize = pageTemplate
-      ? pageSizes.find((ps) => ps.id === pageTemplate.sizeId)
-      : undefined;
-    const options: GcodeOptions = {
-      arcFitting: false,
-      arcTolerance: 0.01,
-      optimisePaths: prefs.optimise,
-      joinPaths: prefs.joinPaths,
-      joinTolerance: prefs.joinTolerance,
-      liftPenAtEnd: prefs.liftPenAtEnd,
-      returnToHome: prefs.returnToHome,
-      customStartGcode: prefs.customStartGcode,
-      customEndGcode: prefs.customEndGcode,
-      pageClip:
-        activePageSize && prefs.clipMode !== "none"
-          ? {
-              widthMM: pageTemplate!.landscape
-                ? activePageSize.heightMM
-                : activePageSize.widthMM,
-              heightMM: pageTemplate!.landscape
-                ? activePageSize.widthMM
-                : activePageSize.heightMM,
-              marginMM:
-                prefs.clipMode === "margin"
-                  ? (pageTemplate!.marginMM ?? 20)
-                  : (prefs.clipOffsetMM ?? 0),
-            }
-          : undefined,
-    };
-
-    // Only process groups that have at least one visible path in them.
-    // Also include a synthetic "ungrouped" entry for imports not in any group.
-    type GroupEntry = {
-      id: string | null;
-      name: string;
-      objects: VectorObject[];
-    };
-    const ungroupedObjects = toVectorObjectsUngrouped();
-    const allEntries: GroupEntry[] = [
-      ...(ungroupedObjects.length > 0
-        ? [{ id: null, name: "ungrouped", objects: ungroupedObjects }]
-        : []),
-      ...layerGroups
-        .map((g) => ({
-          id: g.id,
-          name: g.name,
-          objects: toVectorObjectsForGroup(g.id),
-        }))
-        .filter((e) => e.objects.length > 0),
-    ];
-    if (allEntries.length === 0) return;
-
-    // Ask for a save directory once when the user wants local files.
-    let saveDir: string | null = null;
-    if (prefs.saveLocally) {
-      saveDir = await window.terraForge.fs.chooseDirectory();
-      if (!saveDir) return; // user cancelled
-    }
-
-    setGenerating(true);
-
-    for (const group of allEntries) {
-      const objects = group.objects;
-      const taskId = uuid();
-      const safeName = group.name.replace(/[\\/:*?"<>|]/g, "_");
-      const defaultFilename = prefs.optimise
-        ? `${safeName}_opt.gcode`
-        : `${safeName}.gcode`;
-
-      upsertTask({
-        id: taskId,
-        type: "gcode-generate",
-        label: `Generating ${group.name}…`,
-        progress: 0,
-        status: "running",
-      });
-
-      const gcode = await new Promise<string | null>((resolve) => {
-        const worker = new Worker(
-          new URL("../../../workers/svgWorker.ts", import.meta.url),
-          { type: "module" },
-        );
-        worker.onmessage = (e) => {
-          const msg = e.data;
-          if (msg.type === "progress") {
-            upsertTask({
-              id: taskId,
-              type: "gcode-generate",
-              label: `Generating ${group.name}…`,
-              progress: msg.percent,
-              status: "running",
-            });
-          } else if (msg.type === "complete") {
-            worker.terminate();
-            upsertTask({
-              id: taskId,
-              type: "gcode-generate",
-              label: `${group.name} ready`,
-              progress: 100,
-              status: "completed",
-            });
-            resolve(msg.gcode as string);
-          } else {
-            worker.terminate();
-            upsertTask({
-              id: taskId,
-              type: "gcode-generate",
-              label: `${group.name} failed`,
-              progress: null,
-              status: "error",
-            });
-            resolve(null);
-          }
-        };
-        worker.postMessage({
-          type: "generate",
-          taskId,
-          objects,
-          config: cfg,
-          options,
-        });
-      });
-
-      if (!gcode) continue;
-
-      if (prefs.saveLocally && saveDir) {
-        const savePath = `${saveDir}/${defaultFilename}`;
-        await window.terraForge.fs.writeFile(savePath, gcode);
-      }
-
-      if (prefs.uploadToSd && useMachineStore.getState().connected) {
-        const uploadTaskId = uuid();
-        const remotePath = "/" + defaultFilename;
-        try {
-          await window.terraForge.fluidnc.uploadGcode(
-            uploadTaskId,
-            gcode,
-            remotePath,
-          );
-        } catch {
-          // Upload error surfaced via upload task toast
-        }
-      }
-    }
-
-    setGenerating(false);
-  };
 
   return (
     <header className="flex items-center gap-3 px-4 py-2 bg-panel border-b border-border-ui shrink-0">
@@ -1445,7 +545,7 @@ export function Toolbar({
           onSave={async () => {
             setShowCloseDialog(false);
             await handleSaveLayout();
-            // Only clear canvas once save dialog resolves (user may cancel it)
+            // Only clear canvas once save dialog resolves (user may cancel it).
             // We clear unconditionally because the user explicitly chose Save —
             // if they cancelled the file picker we still dismiss the close dialog.
             clearImports();
@@ -1461,6 +561,7 @@ export function Toolbar({
           message={`The canvas already has ${imports.length} object${imports.length !== 1 ? "s" : ""}. Opening this layout will replace it.\n\nContinue?`}
           confirmLabel="Replace"
           onConfirm={() => {
+            const { loadLayout } = useCanvasStore.getState();
             loadLayout(
               pendingLayout.imports,
               pendingLayout.layerGroups,
