+++ src/renderer/src/features/machine/hooks/useJobActions.ts
@@ -1,0 +1,423 @@
+import { useState, useRef } from "react";
+import { v4 as uuid } from "uuid";
+import { useCanvasStore } from "../../../store/canvasStore";
+import { useTaskStore } from "../../../store/taskStore";
+import { useMachineStore } from "../../../store/machineStore";
+import { parseGcode } from "../../../utils/gcodeParser";
+import { type GcodeOptions, type VectorObject } from "../../../../../types";
+import type { GcodePrefs } from "../../../components/GcodeOptionsDialog";
+
+/** Orchestrates machine connection, disconnection, and G-code generation. */
+export function useJobActions() {
+  const activeConfig = useMachineStore((s) => s.activeConfig);
+  const setConnected = useMachineStore((s) => s.setConnected);
+  const setSelectedJobFile = useMachineStore((s) => s.setSelectedJobFile);
+
+  const imports = useCanvasStore((s) => s.imports);
+  const layerGroups = useCanvasStore((s) => s.layerGroups);
+  const pageTemplate = useCanvasStore((s) => s.pageTemplate);
+  const pageSizes = useCanvasStore((s) => s.pageSizes);
+  const setGcodeToolpath = useCanvasStore((s) => s.setGcodeToolpath);
+  const setGcodeSource = useCanvasStore((s) => s.setGcodeSource);
+  const selectToolpath = useCanvasStore((s) => s.selectToolpath);
+  const toVectorObjectsForGroup = useCanvasStore(
+    (s) => s.toVectorObjectsForGroup,
+  );
+  const toVectorObjectsUngrouped = useCanvasStore(
+    (s) => s.toVectorObjectsUngrouped,
+  );
+
+  const upsertTask = useTaskStore((s) => s.upsertTask);
+  const registerCancelCallback = useTaskStore((s) => s.registerCancelCallback);
+  const unregisterCancelCallback = useTaskStore(
+    (s) => s.unregisterCancelCallback,
+  );
+
+  const [isConnecting, setIsConnecting] = useState(false);
+  const [generating, setGenerating] = useState(false);
+
+  // Holds a reference to the active G-code worker so cancellation can reach it.
+  const activeWorkerRef = useRef<{ worker: Worker; taskId: string } | null>(
+    null,
+  );
+
+  /** Connects to the active machine via Wi-Fi WebSocket or serial port. */
+  const handleConnect = async () => {
+    const cfg = activeConfig();
+    if (!cfg) return;
+    const taskId = uuid();
+    const label =
+      cfg.connection.type === "wifi"
+        ? `Connecting to ${cfg.connection.host}…`
+        : `Connecting to ${cfg.connection.serialPath}…`;
+    upsertTask({
+      id: taskId,
+      type: "ws-connect",
+      label,
+      progress: null,
+      status: "running",
+    });
+    setIsConnecting(true);
+    try {
+      if (cfg.connection.type === "wifi") {
+        await window.terraForge.fluidnc.connectWebSocket(
+          cfg.connection.host!,
+          cfg.connection.port ?? 80,
+          cfg.connection.wsPort,
+        );
+      } else {
+        await window.terraForge.serial.connect(
+          cfg.connection.serialPath!,
+          115200,
+        );
+      }
+      setConnected(true);
+      upsertTask({
+        id: taskId,
+        type: "ws-connect",
+        label: "Connected",
+        progress: 100,
+        status: "completed",
+      });
+    } catch (err) {
+      console.error("Connection failed", err);
+      upsertTask({
+        id: taskId,
+        type: "ws-connect",
+        label: "Connection failed",
+        progress: null,
+        status: "error",
+        error: err instanceof Error ? err.message : String(err),
+      });
+    } finally {
+      setIsConnecting(false);
+    }
+  };
+
+  /** Disconnects from the active machine connection. */
+  const handleDisconnect = async () => {
+    const cfg = activeConfig();
+    if (!cfg) return;
+    if (cfg.connection.type === "wifi") {
+      await window.terraForge.fluidnc.disconnectWebSocket();
+    } else {
+      await window.terraForge.serial.disconnect();
+    }
+    setConnected(false);
+  };
+
+  /** Builds a GcodeOptions object from user preferences and the current page template. */
+  const buildOptions = (prefs: GcodePrefs): GcodeOptions => {
+    const activePageSize = pageTemplate
+      ? pageSizes.find((ps) => ps.id === pageTemplate.sizeId)
+      : undefined;
+    return {
+      arcFitting: false,
+      arcTolerance: 0.01,
+      optimisePaths: prefs.optimise,
+      joinPaths: prefs.joinPaths,
+      joinTolerance: prefs.joinTolerance,
+      liftPenAtEnd: prefs.liftPenAtEnd,
+      returnToHome: prefs.returnToHome,
+      customStartGcode: prefs.customStartGcode,
+      customEndGcode: prefs.customEndGcode,
+      pageClip:
+        activePageSize && prefs.clipMode !== "none"
+          ? {
+              widthMM: pageTemplate!.landscape
+                ? activePageSize.heightMM
+                : activePageSize.widthMM,
+              heightMM: pageTemplate!.landscape
+                ? activePageSize.widthMM
+                : activePageSize.heightMM,
+              marginMM:
+                prefs.clipMode === "margin"
+                  ? (pageTemplate!.marginMM ?? 20)
+                  : (prefs.clipOffsetMM ?? 0),
+            }
+          : undefined,
+    };
+  };
+
+  /**
+   * Generates G-code for all canvas imports as a single file.
+   * Routes to per-group export when `prefs.exportPerGroup` is enabled.
+   */
+  const handleGenerateGcode = async (prefs: GcodePrefs) => {
+    const cfg = activeConfig();
+    if (!cfg || imports.length === 0) return;
+
+    if (prefs.exportPerGroup && layerGroups.length > 0) {
+      await handleGenerateGcodePerGroup(prefs, cfg);
+      return;
+    }
+
+    setGenerating(true);
+    const taskId = uuid();
+    const options = buildOptions(prefs);
+
+    const worker = new Worker(
+      new URL("../../../../../workers/svgWorker.ts", import.meta.url),
+      { type: "module" },
+    );
+
+    activeWorkerRef.current = { worker, taskId };
+
+    registerCancelCallback(taskId, () => {
+      worker.postMessage({ type: "cancel", taskId });
+    });
+
+    upsertTask({
+      id: taskId,
+      type: "gcode-generate",
+      label: "Generating G-code",
+      progress: 0,
+      status: "running",
+    });
+
+    worker.onmessage = async (e) => {
+      const msg = e.data;
+      if (msg.type === "progress") {
+        upsertTask({
+          id: taskId,
+          type: "gcode-generate",
+          label: "Generating G-code",
+          progress: msg.percent,
+          status: "running",
+        });
+      } else if (msg.type === "complete") {
+        worker.terminate();
+        activeWorkerRef.current = null;
+        unregisterCancelCallback(taskId);
+        setGenerating(false);
+
+        const baseName =
+          imports.length === 1
+            ? imports[0].name
+            : `${imports[0].name}+${imports.length - 1}`;
+        const safeName = baseName
+          .replace(/\.[^.]+$/, "")
+          .replace(/[\\/:*?"<>|]/g, "_");
+        const defaultFilename = prefs.optimise
+          ? `${safeName}_opt.gcode`
+          : `${safeName}.gcode`;
+
+        upsertTask({
+          id: taskId,
+          type: "gcode-generate",
+          label: "G-code ready",
+          progress: 100,
+          status: "completed",
+        });
+
+        const toolpath = parseGcode(msg.gcode);
+        setGcodeToolpath(toolpath);
+        selectToolpath(true);
+
+        if (prefs.uploadToSd && useMachineStore.getState().connected) {
+          const uploadTaskId = uuid();
+          const remotePath = "/" + defaultFilename;
+          try {
+            await window.terraForge.fluidnc.uploadGcode(
+              uploadTaskId,
+              msg.gcode,
+              remotePath,
+            );
+            setSelectedJobFile({
+              path: remotePath,
+              source: "sd",
+              name: defaultFilename,
+            });
+            setGcodeSource({
+              path: remotePath,
+              name: defaultFilename,
+              source: "sd",
+            });
+          } catch {
+            // Upload error is already surfaced via the upload task toast
+          }
+        }
+
+        if (prefs.saveLocally) {
+          const savePath =
+            await window.terraForge.fs.saveGcodeDialog(defaultFilename);
+          if (savePath) {
+            await window.terraForge.fs.writeFile(savePath, msg.gcode);
+            setGcodeSource({
+              path: savePath,
+              name: defaultFilename,
+              source: "local",
+            });
+          }
+        }
+      } else if (msg.type === "cancelled") {
+        upsertTask({
+          id: taskId,
+          type: "gcode-generate",
+          label: "G-code cancelled",
+          progress: null,
+          status: "cancelled",
+        });
+        worker.terminate();
+        activeWorkerRef.current = null;
+        unregisterCancelCallback(taskId);
+        setGenerating(false);
+      } else if (msg.type === "error") {
+        upsertTask({
+          id: taskId,
+          type: "gcode-generate",
+          label: "G-code failed",
+          progress: null,
+          status: "error",
+          error: msg.error,
+        });
+        worker.terminate();
+        activeWorkerRef.current = null;
+        unregisterCancelCallback(taskId);
+        setGenerating(false);
+      }
+    };
+
+    worker.postMessage({
+      type: "generate",
+      taskId,
+      objects: useCanvasStore.getState().toVectorObjects(),
+      config: cfg,
+      options,
+    });
+  };
+
+  /**
+   * Generates separate G-code files for each layer group.
+   * Runs workers sequentially — one per group — and saves/uploads each output.
+   */
+  const handleGenerateGcodePerGroup = async (
+    prefs: GcodePrefs,
+    cfg: ReturnType<typeof activeConfig>,
+  ) => {
+    if (!cfg) return;
+    const options = buildOptions(prefs);
+
+    type GroupEntry = {
+      id: string | null;
+      name: string;
+      objects: VectorObject[];
+    };
+    const ungroupedObjects = toVectorObjectsUngrouped();
+    const allEntries: GroupEntry[] = [
+      ...(ungroupedObjects.length > 0
+        ? [{ id: null, name: "ungrouped", objects: ungroupedObjects }]
+        : []),
+      ...layerGroups
+        .map((g) => ({
+          id: g.id,
+          name: g.name,
+          objects: toVectorObjectsForGroup(g.id),
+        }))
+        .filter((e) => e.objects.length > 0),
+    ];
+    if (allEntries.length === 0) return;
+
+    let saveDir: string | null = null;
+    if (prefs.saveLocally) {
+      saveDir = await window.terraForge.fs.chooseDirectory();
+      if (!saveDir) return;
+    }
+
+    setGenerating(true);
+
+    for (const group of allEntries) {
+      const objects = group.objects;
+      const taskId = uuid();
+      const safeName = group.name.replace(/[\\/:*?"<>|]/g, "_");
+      const defaultFilename = prefs.optimise
+        ? `${safeName}_opt.gcode`
+        : `${safeName}.gcode`;
+
+      upsertTask({
+        id: taskId,
+        type: "gcode-generate",
+        label: `Generating ${group.name}…`,
+        progress: 0,
+        status: "running",
+      });
+
+      const gcode = await new Promise<string | null>((resolve) => {
+        const worker = new Worker(
+          new URL("../../../../../workers/svgWorker.ts", import.meta.url),
+          { type: "module" },
+        );
+        worker.onmessage = (e) => {
+          const msg = e.data;
+          if (msg.type === "progress") {
+            upsertTask({
+              id: taskId,
+              type: "gcode-generate",
+              label: `Generating ${group.name}…`,
+              progress: msg.percent,
+              status: "running",
+            });
+          } else if (msg.type === "complete") {
+            worker.terminate();
+            upsertTask({
+              id: taskId,
+              type: "gcode-generate",
+              label: `${group.name} ready`,
+              progress: 100,
+              status: "completed",
+            });
+            resolve(msg.gcode as string);
+          } else {
+            worker.terminate();
+            upsertTask({
+              id: taskId,
+              type: "gcode-generate",
+              label: `${group.name} failed`,
+              progress: null,
+              status: "error",
+            });
+            resolve(null);
+          }
+        };
+        worker.postMessage({
+          type: "generate",
+          taskId,
+          objects,
+          config: cfg,
+          options,
+        });
+      });
+
+      if (!gcode) continue;
+
+      if (prefs.saveLocally && saveDir) {
+        const savePath = `${saveDir}/${defaultFilename}`;
+        await window.terraForge.fs.writeFile(savePath, gcode);
+      }
+
+      if (prefs.uploadToSd && useMachineStore.getState().connected) {
+        const uploadTaskId = uuid();
+        const remotePath = "/" + defaultFilename;
+        try {
+          await window.terraForge.fluidnc.uploadGcode(
+            uploadTaskId,
+            gcode,
+            remotePath,
+          );
+        } catch {
+          // Upload error surfaced via upload task toast
+        }
+      }
+    }
+
+    setGenerating(false);
+  };
+
+  return {
+    handleConnect,
+    handleDisconnect,
+    isConnecting,
+    handleGenerateGcode,
+    generating,
+  };
+}
